requirements.txt
cairosvg
pillow
numpy
opencv-python-headless